research/augm/requirements.txt
tqdm==4.66.2
pillow==10.3.0